Add and fix build of integration tests:
The integration tests did not build due to circular linking dependencies which Rust cannot resolve yet (linking strategy option is not added yet).
See: https://rust-lang.github.io/rfcs/2951-native-link-modifiers.html#support-linkarg--string-in-addition-to-the-modifiers

Also the bindgen comment generation was disabled, because it breaks the build as well.

Adding a build.rs for bda, adds the necessary linking flags in case of a test build.