Don't use reverse topological order for entry point determination.
This is not possible, because for certain graphs (loops in loops), after resolving them
they can have a node which had an incoming edge, but now have non and are at the top of the rev_topo vector.

So the first node in the rev_topological vector does not imply they are an entry node.